Attempt to use a new scheme to avoid data corruption from writing during phase 1.
The idea was to alternate a batch with phase 1 active (and maybe
another phase) with all phases off, and switch between pushing to the
stack (i.e. effectively reading 2 bits from the disk) and pulling from
it instead of reading during phase 1.

Unfortunately the simplest version gets stuck during seeking, probably
because the duty cycle of seeking (50%) is too low.

Fiddling with this can get it over that barrier but then the track
spiral is quite unstable and doesn't seem useful.